Fixed compilation under gcc. Fixed warnings under gcc. Overloaded operator = is called when it's used in variable definition. Fixed member function call of variable that is a reference to a type with member functions. Added error if one of ternary operator ?: result was void. Variable info tree is more error-prone (but not entirely :( ). Added unit tests for overloaded set and modify operators, member function call (for the case above), double in-place array with integer elements, function prototypes, member function call from member function and failure tests. Fixed some failure tests that failed for wrong reasons.

+++ NULLC/nullc.cpp
@@ -47,7 +47,7 @@ void	nullcInit()
 	nullcInitCustomAlloc(NULL, NULL);
 }
 
-void	nullcInitCustomAlloc(void* (NCDECL *allocFunc)(size_t), void (NCDECL *deallocFunc)(void*))
+void	nullcInitCustomAlloc(void* (NCDECL *allocFunc)(int), void (NCDECL *deallocFunc)(void*))
 {
 	NULLC::alloc = allocFunc ? allocFunc : NULLC::defaultAlloc;
 	NULLC::dealloc = deallocFunc ? deallocFunc : NULLC::defaultDealloc;
